Alteracao do tamanho do registrador de aquisicao do ADS de int para long long, pois como o firmware original foi feito para o ARM (registradores de 32bits) dava overflow no Arduino UNO, visto que os registradores são de 16 bits.

+++ Teste_ADS1248_UNO/Teste_ADS1248/Teste_ADS1248.ino
@@ -196,7 +196,9 @@ bool ads1248_ajustar_canal_ganho(int MUX_SP, int MUX_SN, int PGA)
 
 double ads1248_ler(int ganho) 
 {
-  int MSB, midSB, LSB;
+  //int MSB, midSB, LSB;
+  //unsigned long int palavra;
+  long long MSB, midSB, LSB;
   unsigned long int palavra;
   double tensao_ads1248[200];
   double media_tensao_ads1248, media_tensao_ads1248_corrigida;
